requirements.txt
asgiref==3.4.1
Django==3.2.6
django-rest-framework==0.1.0
djangorestframework==3.12.4
pytils==0.3
pytz==2021.1
sqlparse==0.4.1
